Add ManualOptModel
